# frontend/src/types/openapi.json
{"openapi":"3.0.2","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/auth/login":{"post":{"tags":["auth"],"summary":"Auth:Jwt.Login","operationId":"auth_jwt_login_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_auth_jwt_login_auth_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"string"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorModel"},"examples":{"LOGIN_BAD_CREDENTIALS":{"summary":"Bad credentials or the user is inactive.","value":{"detail":"LOGIN_BAD_CREDENTIALS"}},"LOGIN_USER_NOT_VERIFIED":{"summary":"The user is not verified.","value":{"detail":"LOGIN_USER_NOT_VERIFIED"}}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/logout":{"post":{"tags":["auth"],"summary":"Auth:Jwt.Logout","operationId":"auth_jwt_logout_auth_logout_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"string"}}}},"401":{"description":"Missing token or inactive user."}},"security":[{"APIKeyCookie":[]}]}},"/auth/forgot-password":{"post":{"tags":["auth"],"summary":"Reset:Forgot Password","operationId":"reset_forgot_password_auth_forgot_password_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_reset_forgot_password_auth_forgot_password_post"}}},"required":true},"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"string"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/reset-password":{"post":{"tags":["auth"],"summary":"Reset:Reset Password","operationId":"reset_reset_password_auth_reset_password_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_reset_reset_password_auth_reset_password_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"string"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorModel"},"examples":{"RESET_PASSWORD_BAD_TOKEN":{"summary":"Bad or expired token.","value":{"detail":"RESET_PASSWORD_BAD_TOKEN"}},"RESET_PASSWORD_INVALID_PASSWORD":{"summary":"Password validation failed.","value":{"detail":{"code":"RESET_PASSWORD_INVALID_PASSWORD","reason":"Password should be at least 3 characters"}}}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/register":{"post":{"tags":["auth"],"summary":"Register:Register","operationId":"register_register_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"string"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorModel"},"examples":{"REGISTER_USER_ALREADY_EXISTS":{"summary":"A user with this email already exists.","value":{"detail":"REGISTER_USER_ALREADY_EXISTS"}},"REGISTER_INVALID_PASSWORD":{"summary":"Password validation failed.","value":{"detail":{"code":"REGISTER_INVALID_PASSWORD","reason":"Password should beat least 3 characters"}}}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user":{"get":{"tags":["user"],"summary":"Get All Users","operationId":"get_all_users_user_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"string"}}}}},"security":[{"APIKeyCookie":[]}]}},"/user/{user_id}/reset-password":{"patch":{"tags":["user"],"summary":"Reset Password","operationId":"reset_password_user__user_id__reset_password_patch","parameters":[{"required":true,"schema":{"title":"User Id","type":"integer"},"name":"user_id","in":"path"},{"required":false,"schema":{"title":"New Password","type":"string"},"name":"new_password","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"string"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyCookie":[]}]}},"/user/{user_id}/limit":{"post":{"tags":["user"],"summary":"Update Limit","operationId":"update_limit_user__user_id__limit_post","parameters":[{"required":true,"schema":{"title":"User Id","type":"integer"},"name":"user_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LimitSchema"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"string"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyCookie":[]}]}},"/user/me":{"get":{"tags":["user"],"summary":"Users:Current User","operationId":"users_current_user_user_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"string"}}}},"401":{"description":"Missing token or inactive user."}},"security":[{"APIKeyCookie":[]}]},"patch":{"tags":["user"],"summary":"Users:Patch Current User","operationId":"users_patch_current_user_user_me_patch","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"string"}}}},"401":{"description":"Missing token or inactive user."},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorModel"},"examples":{"UPDATE_USER_EMAIL_ALREADY_EXISTS":{"summary":"A user with this email already exists.","value":{"detail":"UPDATE_USER_EMAIL_ALREADY_EXISTS"}},"UPDATE_USER_INVALID_PASSWORD":{"summary":"Password validation failed.","value":{"detail":{"code":"UPDATE_USER_INVALID_PASSWORD","reason":"Password should beat least 3 characters"}}}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyCookie":[]}]}},"/user/{id}":{"get":{"tags":["user"],"summary":"Users:User","operationId":"users_user_user__id__get","parameters":[{"required":true,"schema":{"title":"Id"},"name":"id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"string"}}}},"401":{"description":"Missing token or inactive user."},"403":{"description":"Not a superuser."},"404":{"description":"The user does not exist."},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyCookie":[]}]},"delete":{"tags":["user"],"summary":"Users:Delete User","operationId":"users_delete_user_user__id__delete","parameters":[{"required":true,"schema":{"title":"Id"},"name":"id","in":"path"}],"responses":{"204":{"description":"Successful Response"},"401":{"description":"Missing token or inactive user."},"403":{"description":"Not a superuser."},"404":{"description":"The user does not exist."},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyCookie":[]}]},"patch":{"tags":["user"],"summary":"Users:Patch User","operationId":"users_patch_user_user__id__patch","parameters":[{"required":true,"schema":{"title":"Id"},"name":"id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"string"}}}},"401":{"description":"Missing token or inactive user."},"403":{"description":"Not a superuser."},"404":{"description":"The user does not exist."},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorModel"},"examples":{"UPDATE_USER_EMAIL_ALREADY_EXISTS":{"summary":"A user with this email already exists.","value":{"detail":"UPDATE_USER_EMAIL_ALREADY_EXISTS"}},"UPDATE_USER_INVALID_PASSWORD":{"summary":"Password validation failed.","value":{"detail":{"code":"UPDATE_USER_INVALID_PASSWORD","reason":"Password should beat least 3 characters"}}}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyCookie":[]}]}},"/conv":{"get":{"tags":["conversation"],"summary":"Get All Conversations","description":"对于普通用户，返回其自己的有效会话\n对于管理员，返回所有对话，并可以指定是否只返回有效会话","operationId":"get_all_conversations_conv_get","parameters":[{"required":false,"schema":{"title":"Valid Only","type":"boolean","default":true},"name":"valid_only","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"string"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyCookie":[]}]}},"/conv/{conversation_id}":{"get":{"tags":["conversation"],"summary":"Get Conversation History","operationId":"get_conversation_history_conv__conversation_id__get","parameters":[{"required":true,"schema":{"title":"Conversation Id","type":"string"},"name":"conversation_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"string"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyCookie":[]}]},"delete":{"tags":["conversation"],"summary":"Delete Conversation","operationId":"delete_conversation_conv__conversation_id__delete","parameters":[{"required":true,"schema":{"title":"Conversation Id","type":"string"},"name":"conversation_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"string"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyCookie":[]}]},"patch":{"tags":["conversation"],"summary":"Change Conversation Title","operationId":"change_conversation_title_conv__conversation_id__patch","parameters":[{"required":true,"schema":{"title":"Conversation Id","type":"string"},"name":"conversation_id","in":"path"},{"required":true,"schema":{"title":"Title","type":"string"},"name":"title","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"string"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyCookie":[]}]}},"/conv/{conversation_id}/vanish":{"delete":{"tags":["conversation"],"summary":"Vanish Conversation","operationId":"vanish_conversation_conv__conversation_id__vanish_delete","parameters":[{"required":true,"schema":{"title":"Conversation Id","type":"string"},"name":"conversation_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"string"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyCookie":[]}]}},"/conv/{conversation_id}/assign/{username}":{"patch":{"tags":["conversation"],"summary":"Assign Conversation","operationId":"assign_conversation_conv__conversation_id__assign__username__patch","parameters":[{"required":true,"schema":{"title":"Username","type":"string"},"name":"username","in":"path"},{"required":true,"schema":{"title":"Conversation Id","type":"string"},"name":"conversation_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"string"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyCookie":[]}]}},"/conv/{conversation_id}/gen_title":{"patch":{"tags":["conversation"],"summary":"Generate Conversation Title","operationId":"generate_conversation_title_conv__conversation_id__gen_title_patch","parameters":[{"required":true,"schema":{"title":"Conversation Id","type":"string"},"name":"conversation_id","in":"path"},{"required":true,"schema":{"title":"Message Id","type":"string"},"name":"message_id","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"string"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyCookie":[]}]}},"/status":{"get":{"tags":["status"],"summary":"Get Status","operationId":"get_status_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"string"}}}}},"security":[{"APIKeyCookie":[]}]}}},"components":{"schemas":{"Body_auth_jwt_login_auth_login_post":{"title":"Body_auth_jwt_login_auth_login_post","required":["username","password"],"type":"object","properties":{"grant_type":{"title":"Grant Type","pattern":"password","type":"string"},"username":{"title":"Username","type":"string"},"password":{"title":"Password","type":"string"},"scope":{"title":"Scope","type":"string","default":""},"client_id":{"title":"Client Id","type":"string"},"client_secret":{"title":"Client Secret","type":"string"}}},"Body_reset_forgot_password_auth_forgot_password_post":{"title":"Body_reset_forgot_password_auth_forgot_password_post","required":["email"],"type":"object","properties":{"email":{"title":"Email","type":"string","format":"email"}}},"Body_reset_reset_password_auth_reset_password_post":{"title":"Body_reset_reset_password_auth_reset_password_post","required":["token","password"],"type":"object","properties":{"token":{"title":"Token","type":"string"},"password":{"title":"Password","type":"string"}}},"ChatStatus":{"title":"ChatStatus","enum":["asking","queueing","idling"],"description":"An enumeration."},"ConversationSchema":{"title":"ConversationSchema","type":"object","properties":{"id":{"title":"Id","type":"integer","default":-1},"conversation_id":{"title":"Conversation Id","type":"string","format":"uuid"},"title":{"title":"Title","type":"string"},"user_id":{"title":"User Id","type":"integer"},"is_valid":{"title":"Is Valid","type":"boolean"},"is_public":{"title":"Is Public","type":"boolean"},"use_paid":{"title":"Use Paid","type":"boolean"},"create_time":{"title":"Create Time","type":"string","format":"date-time"},"active_time":{"title":"Active Time","type":"string","format":"date-time"}}},"ErrorModel":{"title":"ErrorModel","required":["detail"],"type":"object","properties":{"detail":{"title":"Detail","anyOf":[{"type":"string"},{"type":"object","additionalProperties":{"type":"string"}}]}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"LimitSchema":{"title":"LimitSchema","type":"object","properties":{"can_use_paid":{"title":"Can Use Paid","type":"boolean"},"max_conv_count":{"title":"Max Conv Count","type":"integer"},"available_ask_count":{"title":"Available Ask Count","type":"integer"}}},"ServerStatusSchema":{"title":"ServerStatusSchema","type":"object","properties":{"active_user_in_5m":{"title":"Active User In 5M","type":"integer"},"active_user_in_1h":{"title":"Active User In 1H","type":"integer"},"active_user_in_1d":{"title":"Active User In 1D","type":"integer"},"is_chatbot_busy":{"title":"Is Chatbot Busy","type":"boolean"},"chatbot_waiting_count":{"title":"Chatbot Waiting Count","type":"integer"}}},"UserCreate":{"title":"UserCreate","required":["email","password","username","nickname"],"type":"object","properties":{"email":{"title":"Email","type":"string"},"password":{"title":"Password","type":"string"},"is_active":{"title":"Is Active","type":"boolean","default":true},"is_superuser":{"title":"Is Superuser","type":"boolean","default":false},"is_verified":{"title":"Is Verified","type":"boolean","default":false},"username":{"title":"Username","type":"string"},"nickname":{"title":"Nickname","type":"string"},"can_use_paid":{"title":"Can Use Paid","type":"boolean","default":false},"max_conv_count":{"title":"Max Conv Count","type":"integer","default":-1},"available_ask_count":{"title":"Available Ask Count","type":"integer","default":-1}}},"UserRead":{"title":"UserRead","required":["id","email","is_active","is_superuser","is_verified","username","nickname","chat_status","can_use_paid"],"type":"object","properties":{"id":{"title":"Id","type":"integer"},"email":{"title":"Email","type":"string"},"is_active":{"title":"Is Active","type":"boolean"},"is_superuser":{"title":"Is Superuser","type":"boolean"},"is_verified":{"title":"Is Verified","type":"boolean"},"username":{"title":"Username","type":"string"},"nickname":{"title":"Nickname","type":"string"},"active_time":{"title":"Active Time","type":"string","format":"date-time"},"chat_status":{"$ref":"#/components/schemas/ChatStatus"},"can_use_paid":{"title":"Can Use Paid","type":"boolean"},"max_conv_count":{"title":"Max Conv Count","type":"integer"},"available_ask_count":{"title":"Available Ask Count","type":"integer"}},"description":"Base User model."},"UserUpdate":{"title":"UserUpdate","required":["nickname"],"type":"object","properties":{"id":{"title":"Id"},"email":{"title":"Email","type":"string"},"is_active":{"title":"Is Active","type":"boolean","default":true},"is_superuser":{"title":"Is Superuser","type":"boolean","default":false},"is_verified":{"title":"Is Verified","type":"boolean","default":false},"nickname":{"title":"Nickname","type":"string"}},"description":"Base User model."},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}},"securitySchemes":{"APIKeyCookie":{"type":"apiKey","in":"cookie","name":"user_auth"}}}}